working on refactor seed file
