+++ server/db/seed/challenge.js
@@ -1,0 +1,29 @@
+const challengesSeed = [
+    {
+      name: 'Make circle',
+      description: 'Make a red circle with radius 100px',
+      difficulty: 1,
+      html: 'dummy data',
+      css: 'dummy data',
+      js: 'dummy data',
+    },
+    {
+      name: 'Make square',
+      description: 'Make a blue square that is 100px wide and high',
+      difficulty: 1,
+    },
+    {
+      name: 'Make rectangle',
+      description: 'Make a yellow rectangle that is 200px wide and 100px high',
+      difficulty: 2,
+    },
+    {
+      name: 'Make rectangle',
+      description: 'Make an orange rectangle that is 100px wide and 200px high',
+      difficulty: 2,
+    },
+  ];
+
+module.exports = {
+    challengesSeed
+}
