+++ server/db/seed/user.js
@@ -1,0 +1,12 @@
+const usersSeed = [
+    { email: 'grant@geezemail.com', password: '1234' },
+    { email: 'kristy@geezemail.com', password: '1234' },
+    { email: 'haoyu@geezemail.com', password: '1234' },
+    { email: 'thee@geezemail.com', password: '1234' },
+    { email: 'a@a.com', password: 'a' },
+    { email: 'b@b.com', password: 'b'},
+  ];
+
+module.exports = {
+    usersSeed
+}
